third_party/blink/web_tests/platform/generic/fast/forms/checkValidity-handler-updates-dom-expected.txt

HTMLFormElement::checkValidity() with cases that event handlers called by checkValidity() updates DOM structure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The target form is removed.
PASS document.getElementById("f1").checkValidity() is false


A control to be checked is removed.
PASS document.getElementById("f1").checkValidity() is false
PASS handler2Called is false


A new control is added.
PASS document.getElementById("f1").checkValidity() is false
PASS handler2Called is false


A control is moved to another form.
PASS document.getElementById("f1").checkValidity() is false
PASS handler2Called is false


A control is moved to another document.
PASS document.getElementById("f1").checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
